my-first-crud-api-BIKErVzMPL/requirements.txt
attrs==22.1.0
aws-cdk-lib==2.39.1
black==22.8.0
cattrs==22.1.0
click==8.1.3
constructs==10.1.102
exceptiongroup==1.0.0rc9
importlib-metadata==4.12.0
jsii==1.67.0
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
publication==0.0.3
python-dateutil==2.8.2
six==1.16.0
tomli==2.0.1
typed-ast==1.5.4
typeguard==2.13.3
typing_extensions==4.3.0
zipp==3.8.1
